% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezknitr-package.R
\docType{package}
\name{ezknitr-package}
\alias{ezknitr}
\alias{ezknitr-package}
\title{ezknitr: Avoid the Typical Working Directory Pain When Using 'knitr'}
\description{
An extension of 'knitr' that adds flexibility in several ways. One common source of frustration with 'knitr' is that it assumes the directory where the source file lives should be the working directory, which is often not true. 'ezknitr' addresses this problem by giving you complete control over where all the inputs and outputs are, and adds several other convenient features to make rendering markdown/HTML documents easier.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://docs.ropensci.org/ezknitr/}
  \item \url{https://github.com/ropensci/ezknitr}
  \item Report bugs at \url{https://github.com/ropensci/ezknitr/issues}
}

}
\author{
\strong{Maintainer}: Dean Attali \email{daattali@gmail.com}

}
\keyword{internal}
