% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezknitr.R
\docType{package}
\name{ezknitr}
\alias{ezknitr}
\alias{ezknitr-package}
\title{ezknitr}
\description{
Avoid the Typical Working Directory Pain When Using 'knitr'
}
\details{
\code{ezknitr} is an extension of \code{knitr} that adds flexibility in several
ways. Most importantly, it allows you to have a complex directory structure
and lets you take complete control over where all your inputs are and where
the outputs should go.
}

