% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ezrender.R
\docType{package}
\name{ezrender}
\alias{ezrender}
\alias{ezrender-package}
\title{ezrender}
\description{
Render Markdown/HTML Reports from R Without the Typical Pain of Dealing with
Working Directories
}

