% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_output_dir.R
\name{open_output_dir}
\alias{open_output_dir}
\title{Open the directory containing the output from the previous ezknitr command}
\usage{
open_output_dir()
}
\description{
Call this function after running an \link[ezknitr]{ezspin} or
\link[ezknitr]{ezknit} command to open the resulting output directory in your
file browser.
}
\examples{
\dontrun{
library(ezknitr)
setup_ezspin_test()
ezspin("R/ezspin_test.R", wd = "ezknitr_test")
open_output_dir()
}
}

