% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_default_params.R
\name{set_default_params}
\alias{set_default_params}
\title{Set default parameters}
\usage{
set_default_params(params)
}
\arguments{
\item{params}{List of parameters.}
}
\description{
Set default parameters
}

