% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/setup_ezspin_test.R
\name{setup_ezspin_test}
\alias{setup_ezspin_test}
\title{Set up a test directory to experiment with \code{ezspin}}
\usage{
setup_ezspin_test()
}
\description{
This function creates a few directories that try to mimic a real
data-analysis project structure, and adds a simple R script and a data file.
After setting up these files and directories, you can run both
\code{knitr::spin} and \code{ezknitr::ezspin} on the R script to see the
benefits of \code{ezspin}.
}
\details{
The console output from this function will give more specific insturctions on
how to interact with this test directory.
}
\examples{
\dontrun{
library(ezknitr)
setup_ezspin_test()
knitr::spin("R/ezspin_test.R")
setup_ezspin_test()
ezspin("R/ezspin_test.R", wd = "ezknitr_test")
setup_ezspin_test()
ezspin("R/ezspin_test.R", wd = "ezknitr_test",
        out_dir = "output", fig_dir = "coolplots")
setwd(origwd)
unlink(tmp, recursive = TRUE, force = TRUE)
}
}
\seealso{
\code{\link[ezknitr]{ezspin}}

\code{\link[knitr]{spin}}
}

