% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ezknit.R
\name{ezknit}
\alias{ezknit}
\title{Create markdown/HTML reports from Rmd files with no hassle}
\usage{
ezknit(file, wd, outDir, figDir, outSuffix, params = list(),
  verbose = FALSE, chunkOpts = list(tidy = FALSE, error = FALSE),
  warn = TRUE)
}
\arguments{
\item{file}{The path to the R script (if \code{wd} is provided, then this
path is relative to \code{wd}).}

\item{wd}{The working directory to be used in the R script. See 'Detailed:
Arguments'.}

\item{outDir}{The output directory (if \code{wd} is provided, then this path
is relative to \code{wd}). Defaults to the directory containing the R script.}

\item{figDir}{The name (or path) of the directory containing the figures
generated for the markdown document. See 'Detailed Arguments'.}

\item{outSuffix}{A suffix to add to the output files, can be used to
differentiate outputs from runs with different parameters. The name of the
output files is the name of the input script appended by \code{outSuffix},
separated by a dash.}

\item{params}{A named list of parameters to be passed on to the R script.
For example, if the script to execute assumes that there is a variable named
\code{DATASET_NAME}, then you can use
\code{params = list('DATASET_NAME' = 'oct10dat')}}

\item{verbose}{If TRUE, then show the progress of knitting the document.}

\item{chunkOpts}{List of chunk options to use. See \code{?knitr::opts_chunk}
for a list of chunk options.}

\item{warn}{If TRUE, then show warnings (recommended to keep this on)}
}
\value{
The path to the output (invisibly).
}
\description{
Similar to \code{\link[ezrender]{ezspin}}, but works as a replacement
to \code{knitr::knit} instead of \code{knitr::spin}. Read the full
documentation at \code{\link[ezrender]{ezspin}}.
}
\seealso{
\code{\link[ezrender]{ezspin}}
}

