% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezknitr.R
\docType{package}
\name{ezknitr}
\alias{ezknitr}
\alias{ezknitr-package}
\title{ezknitr}
\description{
Avoid the Typical Working Directory Pain When Using 'knitr'
}
\details{
\code{ezknitr} is an extension of \code{knitr} that adds flexibility in several
ways. One common source of frustration with \code{knitr} is that it assumes
the directory where the source file lives should be the working directory,
which is often not true. \code{ezknitr} addresses this problem by giving you
complete control over where all the inputs and outputs are, and adds several
other convenient features to make rendering markdown/HTML documents easier.
}

