% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_output_dir.R
\name{open_output_dir}
\alias{open_output_dir}
\title{Open the directory containing the output from the last ezknitr command}
\usage{
open_output_dir()
}
\description{
Call this function after running \link[ezknitr]{ezspin} or
\link[ezknitr]{ezknit} to open the resulting output directory in your
file browser. This is simply a convenience function so that if you want to
see the results you don't need to navigate to the appropriate folder manually.
}
\examples{
\dontrun{
library(ezknitr)
setup_ezspin_test()
ezspin("R/ezspin_test.R", wd = "ezknitr_test")
open_output_dir()
}
}

