% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_ezspin_test.R
\name{setup_ezspin_test}
\alias{setup_ezspin_test}
\title{Set up a test directory to experiment with \code{ezspin}}
\usage{
setup_ezspin_test()
}
\value{
The path to the root of the test directory.
}
\description{
This function creates a few directories that try to mimic a real
data-analysis project structure, and adds a simple R script and a data file.
After setting up these files and directories, you can run both
\code{knitr::spin} and \code{ezknitr::ezspin} on the R script to see the
benefits of \code{ezspin}.
}
\details{
The console output from this function will give more specific insturctions on
how to interact with this test directory.
}
\examples{
\dontrun{
library(ezknitr)

# setup the test directory structures and run naive spin
setup_ezspin_test()
knitr::spin("ezknitr_test/R/ezspin_test.R")
file.remove(c("ezspin_test.md", "ezspin_test.html"))

# setup the test directory structures and run simple ezspin
setup_ezspin_test()
ezspin("R/ezspin_test.R", wd = "ezknitr_test")

# setup the test directory structures and run ezspin with more parameters
tmp <- setup_ezspin_test()
ezspin("R/ezspin_test.R", wd = "ezknitr_test",
        out_dir = "output", fig_dir = "coolplots")
unlink(tmp, recursive = TRUE, force = TRUE)
}
}
\seealso{
\code{\link[ezknitr]{ezspin}}

\code{\link[knitr]{spin}}
}

